requirements.txt
apache-airflow[postgres,crypto,amazon]==2.3.0 --constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.3.0/constraints-3.9.txt
apache-airflow-providers-apache-livy==2.2.3
pandas==1.4.2
